+++ ddth-dao-core/src/main/java/com/github/ddth/dao/jdbc/utils/IFilter.java
@@ -8,7 +8,7 @@ package com.github.ddth.dao.jdbc.utils;
  */
 public interface IFilter {
     /**
-     * Build the "filter" part (without the WHERE/HAVING/etc keyword), with placeholders and list of values for binding.
+     * Build the "filter" part (without the WHERE/HAVING/etc keyword), with placeholders and list of corresponding binding values.
      *
      * @return
      */